requirements.txt
python-binance==1.0.28
pandas==1.5.3
numpy==1.24.2
matplotlib==3.7.1
schedule==1.1.0
python-dotenv==1.1.0
requests==2.28.2
python-telegram-bot==13.15
ta==0.10.2
websocket-client==1.5.1
scipy==1.10.1
pytz==2023.3
tqdm==4.65.0 